pyproject.toml
[project]
name = "db-design-copilot-api"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "python-dotenv>=1.0",
  "psycopg[binary,pool]>=3.2",
  "pydantic>=2.8",
  "numpy>=1.26",
  "tiktoken>=0.7",
  "PyYAML>=6.0",
  "orjson>=3.10",
  "jinja2>=3.1",
  "httpx>=0.27",
]

[tool.pytest.ini_options]
pythonpath = ["apps/api"]
